backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
websockets==13.0
google-genai==0.8.0
google-cloud-aiplatform==1.71.0
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-adk==0.4.0
python-dotenv==1.0.1
pydantic==2.9.2
pillow==11.0.0
numpy==2.1.3
httpx==0.27.2
python-multipart==0.0.12
